# docs/openapi.json
{
  "components": {
    "schemas": {
      "AgentRunRecord": {
        "additionalProperties": false,
        "properties": {
          "agent_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Agent Name",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "estimated_cost_usd": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Usd"
          },
          "fallback_outcome": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Outcome"
          },
          "input_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "model_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "output_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "record_id": {
            "format": "uuid",
            "title": "Record Id",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "runtime_mode": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Mode"
          },
          "schema_version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "source_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStage"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "tool_activity": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Tool Activity",
            "type": "array"
          },
          "total_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          },
          "trace_id": {
            "anyOf": [
              {
                "maxLength": 300,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "run_id",
          "stage",
          "agent_name",
          "status",
          "started_at"
        ],
        "title": "AgentRunRecord",
        "type": "object"
      },
...
      "CanonicalProduct": {
        "additionalProperties": false,
        "properties": {
          "brand": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "category": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "ean": {
            "anyOf": [
              {
                "maxLength": 14,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ean"
          },
          "listing_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Listing Ids",
            "type": "array"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "product_id": {
            "format": "uuid",
            "title": "Product Id",
            "type": "string"
          },
          "schema_version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "sku": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sku"
          },
          "source_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "upc": {
            "anyOf": [
              {
                "maxLength": 14,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upc"
          }
        },
        "required": [
          "name"
        ],
        "title": "CanonicalProduct",
        "type": "object"
      },
      "CategoryAnalysis": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Category",
            "type": "string"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "evidence_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "minItems": 1,
            "title": "Evidence Ids",
            "type": "array"
          },
          "fit_summary": {
            "maxLength": 1500,
            "minLength": 1,
            "title": "Fit Summary",
            "type": "string"
          },
          "listing_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Listing Ids",
            "type": "array"
          },
          "product_id": {
            "format": "uuid",
            "title": "Product Id",
            "type": "string"
          },
          "schema_version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "source_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "title": "Strengths",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "weaknesses": {
            "items": {
              "type": "string"
            },
            "title": "Weaknesses",
            "type": "array"
          }
        },
        "required": [
          "product_id",
          "category",
          "fit_summary",
          "confidence",
          "evidence_ids"
        ],
        "title": "CategoryAnalysis",
        "type": "object"
      },
...
      "ComparisonMatrix": {
        "additionalProperties": false,
        "properties": {
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/ComparisonCriterion"
            },
            "title": "Criteria",
            "type": "array"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ComparisonRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "schema_version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "title": "ComparisonMatrix",
        "type": "object"
      },
...
      "ConsideredProductOutcomeResponse": {
        "additionalProperties": false,
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/UserAddedProduct"
          },
          "exclusion_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclusion Reason"
          },
          "status": {
            "enum": [
              "confirmed",
              "possible",
              "unresolved",
              "manual",
              "excluded"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "candidate",
          "status"
        ],
        "title": "ConsideredProductOutcomeResponse",
        "type": "object"
      },
...
      "DecisionHistoryResponse": {
        "additionalProperties": false,
        "properties": {
          "original_query": {
            "title": "Original Query",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/DecisionHistoryEntry"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "original_query",
          "versions"
        ],
        "title": "DecisionHistoryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProductListing": {
        "additionalProperties": false,
        "properties": {
          "canonical_url": {
            "anyOf": [
              {
                "format": "uri",
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "captured_at": {
            "format": "date-time",
            "title": "Captured At",
            "type": "string"
          },
          "ean": {
            "anyOf": [
              {
                "maxLength": 14,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ean"
          },
          "listing_id": {
            "format": "uuid",
            "title": "Listing Id",
            "type": "string"
          },
          "price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "product_id": {
            "format": "uuid",
            "title": "Product Id",
            "type": "string"
          },
          "region_availability": {
            "items": {
              "$ref": "#/components/schemas/RegionAvailability"
            },
            "title": "Region Availability",
            "type": "array"
          },
          "retailer_id": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retailer Id"
          },
          "schema_version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "seller": {
            "$ref": "#/components/schemas/SellerProfile"
          },
          "sku": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sku"
          },
          "source_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "minItems": 1,
            "title": "Source Ids",
            "type": "array"
          },
          "source_quality": {
            "$ref": "#/components/schemas/SourceQuality"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "upc": {
            "anyOf": [
              {
                "maxLength": 14,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upc"
          },
          "url": {
            "format": "uri",
            "minLength": 1,
            "title": "Url",
            "type": "string"
          },
          "user_added_matches": {
            "items": {
              "$ref": "#/components/schemas/UserAddedListingMatch"
            },
            "title": "User Added Matches",
            "type": "array"
          }
        },
        "required": [
          "product_id",
          "title",
          "url",
          "seller",
          "source_ids"
        ],
        "title": "ProductListing",
        "type": "object"
      },
...
      "SessionResultsResponse": {
        "additionalProperties": false,
        "properties": {
          "agent_records": {
            "items": {
              "$ref": "#/components/schemas/AgentRunRecord"
            },
            "title": "Agent Records",
            "type": "array"
          },
          "category_analyses": {
            "items": {
              "$ref": "#/components/schemas/CategoryAnalysis"
            },
            "title": "Category Analyses",
            "type": "array"
          },
          "comparison_matrix": {
            "$ref": "#/components/schemas/ComparisonMatrix"
          },
          "considered_products": {
            "items": {
              "$ref": "#/components/schemas/ConsideredProductOutcomeResponse"
            },
            "title": "Considered Products",
            "type": "array"
          },
          "listings": {
            "items": {
              "$ref": "#/components/schemas/ProductListing"
            },
            "title": "Listings",
            "type": "array"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/CanonicalProduct"
            },
            "title": "Products",
            "type": "array"
          },
          "recommendation_bundle": {
            "$ref": "#/components/schemas/RecommendationBundle"
          },
          "result_version": {
            "$ref": "#/components/schemas/ResultVersionResponse"
          },
          "shortlist": {
            "items": {
              "$ref": "#/components/schemas/ShortlistItemResponse"
            },
            "title": "Shortlist",
            "type": "array"
          },
          "source_evidence": {
            "items": {
              "$ref": "#/components/schemas/SourceEvidence"
            },
            "title": "Source Evidence",
            "type": "array"
          },
          "source_snapshots": {
            "items": {
              "$ref": "#/components/schemas/SourceSnapshot"
            },
            "title": "Source Snapshots",
            "type": "array"
          },
          "trust_assessments": {
            "items": {
              "$ref": "#/components/schemas/ListingTrustAssessment"
            },
            "title": "Trust Assessments",
            "type": "array"
          }
        },
        "required": [
          "result_version",
          "trust_assessments",
          "category_analyses",
          "agent_records",
          "comparison_matrix",
          "recommendation_bundle",
          "products",
          "listings",
          "shortlist",
          "considered_products",
          "source_snapshots",
          "source_evidence"
        ],
        "title": "SessionResultsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CartCart Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/sessions/{session_id}/results": {
      "get": {
        "operationId": "get_session_results_api_sessions__session_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResultsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session Results",
        "tags": [
          "results"
        ]
      }
    },
    "/api/sessions/{session_id}/results/history": {
      "get": {
        "operationId": "get_decision_history_api_sessions__session_id__results_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Decision History",
        "tags": [
          "results"
        ]
      }
    },
    "/api/sessions/{session_id}/results/{result_version_id}": {
      "get": {
        "operationId": "get_result_version_api_sessions__session_id__results__result_version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "result_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Result Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResultsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Result Version",
        "tags": [
          "results"
        ]
      }
    },
...
  }
}